# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MaintenancePizza API",
    "version": "1.0.0",
    "description": "Ticket system microservice for store maintenance. An authenticated user files a ticket describing the issues in a store; dispatchers then drive each issue through its lifecycle (diagnosis, technician assignment, attendance, parts, pay/driving, warranty) until completion.\n\nKey design notes:\n- All endpoints require a Sanctum bearer token (`Authorization: Bearer <token>`). There are no login/registration endpoints in this service — tokens are issued elsewhere.\n- Most endpoints are store-scoped under `/api/stores/{store}`, where `{store}` is the store NUMBER (e.g. `03795-00001`), not a numeric id.\n- The system is append-only: there are no update (PUT/PATCH) endpoints. Corrections are made by marking leaf records `mistaken` and creating new ones. Lifecycle changes are explicit POST actions.\n- A ticket's status is DERIVED from its issues and is never set directly.\n- Workflow records (diagnosis, attendance, parts, pay, warranty, assignment, technicians) attach to ONE OR MANY issues via `ticket_issue_ids`, enabling flexible grouping.\n- Attachments are uploaded as multipart `files[]` together with the record they belong to and stored on the public disk; responses include a full `url`."
  },
  "servers": [
    { "url": "http://localhost:8000/api", "description": "Local dev server" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/issues": {
      "get": {
        "tags": ["Issues"], "summary": "List issues", "operationId": "listIssues",
        "parameters": [
          { "$ref": "#/components/parameters/Trashed" },
          { "$ref": "#/components/parameters/PerPage" }
        ],
        "responses": { "200": { "description": "Paginated issues", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IssueCollection" } } } }, "401": { "$ref": "#/components/responses/Unauthenticated" } }
      },
      "post": {
        "tags": ["Issues"], "summary": "Create issue", "operationId": "createIssue",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["title"], "properties": { "title": { "type": "string", "maxLength": 255 }, "description": { "type": "string", "nullable": true } } } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IssueResource" } } } }, "422": { "$ref": "#/components/responses/ValidationError" } }
      }
    },
    "/issues/{issue}": {
      "delete": {
        "tags": ["Issues"], "summary": "Soft-delete issue", "operationId": "deleteIssue",
        "parameters": [{ "name": "issue", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Soft-deleted" } }
      }
    },
    "/issues/{issue}/restore": {
      "post": {
        "tags": ["Issues"], "summary": "Restore soft-deleted issue", "operationId": "restoreIssue",
        "parameters": [{ "name": "issue", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Restored", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IssueResource" } } } } }
      }
    },
    "/technicians": {
      "get": {
        "tags": ["Technicians"], "summary": "List technicians", "operationId": "listTechnicians",
        "parameters": [{ "$ref": "#/components/parameters/Trashed" }, { "$ref": "#/components/parameters/PerPage" }],
        "responses": { "200": { "description": "Paginated technicians" } }
      },
      "post": {
        "tags": ["Technicians"], "summary": "Create technician", "operationId": "createTechnician",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "category_id": { "type": "integer", "nullable": true } } } } } },
        "responses": { "201": { "description": "Created" }, "422": { "$ref": "#/components/responses/ValidationError" } }
      }
    },
...
    "/stores/{store}/tickets/{ticket}/diagnoses": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Add a diagnosis (multipart, optional files)", "operationId": "createDiagnosis",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["ticket_issue_ids"], "properties": { "ticket_issue_ids[]": { "type": "array", "items": { "type": "integer" } }, "body": { "type": "string", "nullable": true }, "files[]": { "type": "array", "items": { "type": "string", "format": "binary" } } } } } } },
        "responses": { "201": { "description": "Diagnosis", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DiagnosisResource" } } } } }
      }
    },
    "/stores/{store}/tickets/{ticket}/diagnoses/{diagnosis}/mistaken": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Mark a diagnosis as mistaken", "operationId": "markDiagnosisMistaken",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }, { "name": "diagnosis", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Diagnosis (mistaken=true)" } }
      }
    },
...
    "/stores/{store}/tickets/{ticket}/attendance-entries/{attendanceEntry}/mistaken": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Mark an attendance entry as mistaken", "operationId": "markAttendanceMistaken",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }, { "name": "attendanceEntry", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Attendance entry (mistaken=true)" } }
      }
    },
    "/stores/{store}/tickets/{ticket}/part-usages": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Add a part usage (multipart, optional files)", "operationId": "createPartUsage",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "required": ["ticket_issue_ids", "part_id", "cost"], "properties": { "ticket_issue_ids[]": { "type": "array", "items": { "type": "integer" } }, "part_id": { "type": "integer" }, "cost": { "type": "number" }, "files[]": { "type": "array", "items": { "type": "string", "format": "binary" } } } } } } },
        "responses": { "201": { "description": "Part usage", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PartUsageResource" } } } } }
      }
    },
    "/stores/{store}/tickets/{ticket}/part-usages/{partUsage}/mistaken": {
      "post": {
        "tags": ["Issue Workflow"], "summary": "Mark a part usage as mistaken", "operationId": "markPartUsageMistaken",
        "parameters": [{ "$ref": "#/components/parameters/Store" }, { "$ref": "#/components/parameters/Ticket" }, { "name": "partUsage", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Part usage (mistaken=true)" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Sanctum personal access token." }
    },
    "parameters": {
      "Store": { "name": "store", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Store number, e.g. 03795-00001", "example": "03795-00001" },
      "Ticket": { "name": "ticket", "in": "path", "required": true, "schema": { "type": "integer" } },
...
      "Trashed": { "name": "trashed", "in": "query", "schema": { "type": "string", "enum": ["with", "only"] }, "description": "Include ('with') or limit to ('only') soft-deleted records" },
      "PerPage": { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 15 } },
...
    },
    "responses": {
      "Unauthenticated": { "description": "Missing/invalid token", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string", "example": "Unauthenticated." } } } } } },
      "ValidationError": { "description": "Validation failed", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "errors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } } } } } } }
    },
    "schemas": {
...
      "IssueResource": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Issue" } } },
      "Issue": { "type": "object", "properties": { "id": { "type": "integer" }, "title": { "type": "string" }, "description": { "type": "string", "nullable": true }, "created_by": { "type": "integer", "nullable": true }, "created_at": { "type": "string", "format": "date-time" }, "deleted_at": { "type": "string", "format": "date-time", "nullable": true } } },
      "IssueCollection": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Issue" } }, "links": { "type": "object" }, "meta": { "type": "object" } } },
...
      "Part": { "type": "object", "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "deleted_at": { "type": "string", "nullable": true } } },
      "Attachment": { "type": "object", "properties": { "id": { "type": "integer" }, "path": { "type": "string" }, "url": { "type": "string" }, "original_name": { "type": "string", "nullable": true }, "mime_type": { "type": "string", "nullable": true }, "size": { "type": "integer", "nullable": true } } },
...
      "Diagnosis": { "type": "object", "properties": { "id": { "type": "integer" }, "body": { "type": "string", "nullable": true }, "mistaken": { "type": "boolean" }, "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } }, "ticket_issue_ids": { "type": "array", "items": { "type": "integer" } } } },
      "DiagnosisResource": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Diagnosis" } } },
...
      "PartUsage": { "type": "object", "properties": { "id": { "type": "integer" }, "part_id": { "type": "integer" }, "part": { "$ref": "#/components/schemas/Part" }, "cost": { "type": "string" }, "mistaken": { "type": "boolean" }, "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } } } },
      "PartUsageResource": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/PartUsage" } } },
...
    }
  }
}